\name{Gini-package}
\alias{Gini-package}
\alias{Gini}
\docType{package}
\title{
Calculates Gini coefficients for regions
}
\description{
Calculates Gini coefficients for regions
}
\details{
\tabular{ll}{
Package: \tab Gini\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2015-06-24\cr
License: \tab GLP-3\cr
}
Functions: huvud, kumul & testdata
}
\author{
Erling Haggstrom Lundevaller

erling.lundevaller@gmail.com
}
\references{

}

\keyword{ package }
\seealso{

}
\examples{

}
