% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gini.R
\name{huvud}
\alias{huvud}
\title{huvud function
This functions takes a data.frame with region and income and calculates Gini coefficients for every region}
\usage{
huvud(data = data)
}
\arguments{
\item{data}{A data.frame with one column region and one column income}
}
\value{
reg.gini A data.frame with columns: reg      gini      mean   quart1    median   quart3 no.cases
}
\description{
huvud function
This functions takes a data.frame with region and income and calculates Gini coefficients for every region
}
\details{
This functions takes a data.frame with region and income and calculates Gini coefficients for every region.
 Note that there must be collumns named region and income
}
\author{
Erling Haggstrom Lundevaller
}

