% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gini.R
\name{kumul}
\alias{kumul}
\title{kumul function
This function makes a cumulative vector}
\usage{
kumul(vec)
}
\arguments{
\item{vec}{a sorted vector with incomes}
}
\value{
kum.vec A cumulative vector of the same length as the input vec
}
\description{
kumul function
This function makes a cumulative vector
}
\details{
Makes a cumulative vector
}
\author{
Erling Haggstrom Lundevaller
}

