% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gini.R
\name{testdata}
\alias{testdata}
\title{Creating data to test the Gini-function}
\usage{
testdata()
}
\value{
A data.frame with test data for the function huvud that calculates gini
}
\description{
Creating data to test the Gini-function
}
\details{
Just for testing!
}

